Refactored SD card handling and acquisition task for improved efficiency. Removed unnecessary mutex usage, updated function parameters, and adjusted buffer handling.

+++ main/sci_scientisst.h
@@ -66,8 +66,8 @@ typedef struct
     uint8_t *frame_buffer[NUM_BUFFERS];           ///< Buffer that holds the frames to be sent
     uint32_t frame_buffer_length_bytes;           ///< Length of each send buffer, set to optimal value depending on com mode
     uint16_t frame_buffer_write_idx[NUM_BUFFERS]; ///< The index of the first free element in for each buffer
-    volatile uint8_t frame_buffer_ready_to_send[NUM_BUFFERS]; ///< If element 0 is set to 1, bt task has to send snd_buff[0]
-    uint8_t tx_curr_buff;                                     ///< Index of the buffer that tx task is currently sending
+    volatile uint16_t frame_buffer_ready_to_send[NUM_BUFFERS]; ///< If element 0 is set to 1, bt task has to send snd_buff[0]
+    uint8_t tx_curr_buff;                                      ///< Index of the buffer that tx task is currently sending
     uint8_t acq_curr_buff;   ///< Index of the buffer that acquisition task is currently using
     uint8_t packet_size;     ///< Current packet size (dependent on number of channels used)
     uint32_t send_threshold; ///< Based on buffer and packet sizes, threshold that marks the buffer as ready to send and
